moved auth api method inside sign api
