+++ example/dapp/lib/widgets/auth_item.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:walletconnect_flutter_v2/walletconnect_flutter_v2.dart';
 import 'package:walletconnect_flutter_v2_dapp/utils/constants.dart';
+
+import 'package:walletconnect_flutter_v2/apis/sign_api/models/auth/auth_client_models.dart';
 
 class AuthItem extends StatelessWidget {
   const AuthItem({
