+++ example/wallet/lib/widgets/wc_connection_request/wc_auth_request_model.dart
@@ -1,4 +1,4 @@
-import 'package:walletconnect_flutter_v2/walletconnect_flutter_v2.dart';
+import 'package:walletconnect_flutter_v2/apis/sign_api/models/auth/auth_client_events.dart';
 
 class WCAuthRequestModel {
   final String iss;
