+++ lib/apis/auth_api/utils/address_utils.dart
@@ -1,13 +1,13 @@
-class AddressUtils {
-  static String getDidAddress(String iss) {
-    return iss.split(':').last;
-  }
+// class AddressUtils {
+//   static String getDidAddress(String iss) {
+//     return iss.split(':').last;
+//   }
 
-  static String getDidChainId(String iss) {
-    return iss.split(':')[3];
-  }
+//   static String getDidChainId(String iss) {
+//     return iss.split(':')[3];
+//   }
 
-  static String getNamespaceDidChainId(String iss) {
-    return iss.substring(iss.indexOf(RegExp(r':')) + 1);
-  }
-}
+//   static String getNamespaceDidChainId(String iss) {
+//     return iss.substring(iss.indexOf(RegExp(r':')) + 1);
+//   }
+// }
