+++ lib/apis/auth_api/utils/auth_constants.dart
@@ -1,10 +1,10 @@
-import 'package:walletconnect_flutter_v2/apis/utils/constants.dart';
+// import 'package:walletconnect_flutter_v2/apis/utils/constants.dart';
 
-class AuthConstants {
-  static const AUTH_REQUEST_EXPIRY_MIN = WalletConnectConstants.FIVE_MINUTES;
-  static const AUTH_REQUEST_EXPIRY_MAX = WalletConnectConstants.SEVEN_DAYS;
+// class AuthConstants {
+//   static const AUTH_REQUEST_EXPIRY_MIN = WalletConnectConstants.FIVE_MINUTES;
+//   static const AUTH_REQUEST_EXPIRY_MAX = WalletConnectConstants.SEVEN_DAYS;
 
-  static const AUTH_DEFAULT_URL = 'https://rpc.walletconnect.com/v1';
+//   static const AUTH_DEFAULT_URL = 'https://rpc.walletconnect.com/v1';
 
-  static const AUTH_CLIENT_PUBLIC_KEY_NAME = 'PUB_KEY';
-}
+//   static const AUTH_CLIENT_PUBLIC_KEY_NAME = 'PUB_KEY';
+// }
