+++ lib/apis/auth_api/utils/auth_utils.dart
@@ -1,5 +1,5 @@
-class AuthUtils {
-  static String generateNonce() {
-    return DateTime.now().millisecondsSinceEpoch.toString();
-  }
-}
+// class AuthUtils {
+//   static String generateNonce() {
+//     return DateTime.now().millisecondsSinceEpoch.toString();
+//   }
+// }
