+++ lib/apis/sign_api/i_sign_engine_common.dart
@@ -8,6 +8,7 @@ import 'package:walletconnect_flutter_v2/apis/sign_api/i_sessions.dart';
 import 'package:walletconnect_flutter_v2/apis/sign_api/models/proposal_models.dart';
 import 'package:walletconnect_flutter_v2/apis/sign_api/models/session_models.dart';
 import 'package:walletconnect_flutter_v2/apis/sign_api/models/sign_client_events.dart';
+import 'package:walletconnect_flutter_v2/apis/sign_api/models/auth/auth_client_models.dart';
 
 abstract class ISignEngineCommon {
   abstract final Event<SessionConnect> onSessionConnect;
@@ -22,6 +23,12 @@ abstract class ISignEngineCommon {
   abstract final ISessions sessions;
   abstract final IGenericStore<SessionRequest> pendingRequests;
 
+  // FORMER AUTH ENGINE PROPERTY
+  abstract final IPairingStore pairings;
+  abstract final IGenericStore<AuthPublicKey> authKeys;
+  abstract final IGenericStore<String> pairingTopics;
+  abstract final IGenericStore<StoredCacao> completeRequests;
+
   Future<void> init();
   Future<void> disconnectSession({
     required String topic,
@@ -33,5 +40,13 @@ abstract class ISignEngineCommon {
   });
   Map<String, ProposalData> getPendingSessionProposals();
 
-  abstract final IPairingStore pairings;
+  // FORMER AUTH ENGINE PROPERTY
+  String formatAuthMessage({
+    required String iss,
+    required CacaoRequestPayload cacaoPayload,
+  });
+  // FORMER AUTH ENGINE PROPERTY
+  Map<int, StoredCacao> getCompletedRequestsForPairing({
+    required String pairingTopic,
+  });
 }
