+++ lib/apis/sign_api/utils/auth/auth_utils.dart
@@ -1,0 +1,5 @@
+class AuthUtils {
+  static String generateNonce() {
+    return DateTime.now().millisecondsSinceEpoch.toString();
+  }
+}
