+++ lib/apis/web3wallet/i_web3wallet.dart
@@ -1,11 +1,9 @@
-import 'package:walletconnect_flutter_v2/apis/auth_api/i_auth_engine_wallet.dart';
 import 'package:walletconnect_flutter_v2/apis/sign_api/i_sign_engine_wallet.dart';
 import 'package:walletconnect_flutter_v2/walletconnect_flutter_v2.dart';
 
-abstract class IWeb3Wallet implements ISignEngineWallet, IAuthEngineWallet {
+abstract class IWeb3Wallet implements ISignEngineWallet {
   final String protocol = 'wc';
   final int version = 2;
 
   abstract final ISignEngine signEngine;
-  abstract final IAuthEngine authEngine;
 }
