+++ test/auth_api/address_utils_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:walletconnect_flutter_v2/walletconnect_flutter_v2.dart';
+import 'package:walletconnect_flutter_v2/apis/sign_api/utils/auth/address_utils.dart';
 
 import '../shared/shared_test_values.dart';
 import 'utils/signature_constants.dart';
