+++ test/auth_api/utils/signature_constants.dart
@@ -1,4 +1,4 @@
-import 'package:walletconnect_flutter_v2/walletconnect_flutter_v2.dart';
+import 'package:walletconnect_flutter_v2/apis/sign_api/models/auth/auth_client_models.dart';
 
 import '../../shared/shared_test_values.dart';
